create ModelUtility
