+++ include/private/ModelUtility.h
@@ -2,7 +2,6 @@
 Copyright (c) 2012-2017 Ben Croston ben@croston.org.
 Copyright (c) 2019, NVIDIA CORPORATION.
 Copyright (c) 2019 Jueon Park(pjueon) bluegbg@gmail.com.
-Copyright (c) 2021 Adam Rasburn blackforestcheesecake@protonmail.ch
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -23,44 +22,22 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 */
 
-#include "private/Model.h"
-#include "private/PythonFunctions.h"
-#include <stdexcept>
+#pragma once
+#ifndef MODEL_UTILITY_H
+#define MODEL_UTILITY_H
+
+#include <string>
 
 namespace GPIO
 {
-    constexpr auto number_of_models = static_cast<int>(sizeof(MODEL_NAMES) / sizeof(Model));
+    // enum
+    enum class Model;
 
-    std::string model_name(Model model)
-    {
-        int idx = static_cast<int>(model);
-        if (idx < 0 || idx >= number_of_models)
-            throw std::runtime_error("model_name error");
-
-        return MODEL_NAMES[idx];
-    }
-
-    int model_name_index(const std::string& name)
-    {
-        auto _name = strip(name);
-
-        for (int idx = 0; idx < number_of_models; idx++)
-        {
-            if (_name == MODEL_NAMES[idx])
-                return idx;
-        }
-
-        return None;
-    }
-
-    Model index_to_model(int idx)
-    {
-        if (idx < 0 || idx >= number_of_models)
-            throw std::runtime_error("index_to_model error");
-
-        return static_cast<Model>(idx);
-    }
-
-    Model name_to_model(const std::string& name) { return index_to_model(model_name_index(name)); }
+    std::string model_name(Model model);
+    int model_name_index(const std::string& name);
+    Model index_to_model(int idx);
+    Model name_to_model(const std::string& name);
 
 } // namespace GPIO
+
+#endif
