+++ src/ModelUtility.cpp
@@ -2,6 +2,7 @@
 Copyright (c) 2012-2017 Ben Croston ben@croston.org.
 Copyright (c) 2019, NVIDIA CORPORATION.
 Copyright (c) 2019 Jueon Park(pjueon) bluegbg@gmail.com.
+Copyright (c) 2021 Adam Rasburn blackforestcheesecake@protonmail.ch
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -22,40 +23,45 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 */
 
-#pragma once
-#ifndef MODEL_H
-#define MODEL_H
-
-#include <string>
+#include "private/Model.h"
+#include "private/ModelUtility.h"
+#include "private/PythonFunctions.h"
+#include <stdexcept>
 
 namespace GPIO
 {
-    // enum
-    enum class Model
+    constexpr auto number_of_models = static_cast<int>(sizeof(MODEL_NAMES) / sizeof(Model));
+
+    std::string model_name(Model model)
     {
-        CLARA_AGX_XAVIER,
-        JETSON_NX,
-        JETSON_XAVIER,
-        JETSON_TX2,
-        JETSON_TX1,
-        JETSON_NANO,
-        JETSON_TX2_NX,
-        JETSON_ORIN
-    };
+        int idx = static_cast<int>(model);
+        if (idx < 0 || idx >= number_of_models)
+            throw std::runtime_error("model_name error");
 
-    // names
-    constexpr const char* MODEL_NAMES[] = {"CLARA_AGX_XAVIER", "JETSON_NX",   "JETSON_XAVIER", "JETSON_TX2",
-                                           "JETSON_TX1",       "JETSON_NANO", "JETSON_TX2_NX", "JETSON_ORIN"};
+        return MODEL_NAMES[idx];
+    }
 
-    // alias
-    constexpr Model CLARA_AGX_XAVIER = Model::CLARA_AGX_XAVIER;
-    constexpr Model JETSON_NX = Model::JETSON_NX;
-    constexpr Model JETSON_XAVIER = Model::JETSON_XAVIER;
-    constexpr Model JETSON_TX2 = Model::JETSON_TX2;
-    constexpr Model JETSON_TX1 = Model::JETSON_TX1;
-    constexpr Model JETSON_NANO = Model::JETSON_NANO;
-    constexpr Model JETSON_TX2_NX = Model::JETSON_TX2_NX;
-    constexpr Model JETSON_ORIN = Model::JETSON_ORIN;
+    int model_name_index(const std::string& name)
+    {
+        auto _name = strip(name);
+
+        for (int idx = 0; idx < number_of_models; idx++)
+        {
+            if (_name == MODEL_NAMES[idx])
+                return idx;
+        }
+
+        return None;
+    }
+
+    Model index_to_model(int idx)
+    {
+        if (idx < 0 || idx >= number_of_models)
+            throw std::runtime_error("index_to_model error");
+
+        return static_cast<Model>(idx);
+    }
+
+    Model name_to_model(const std::string& name) { return index_to_model(model_name_index(name)); }
+
 } // namespace GPIO
-
-#endif
